01_back_end/pyproject.toml
[project]
name = "00-labs"
version = "0.1.0"
description = "This is a collection of labs that showcases how to create deep research agents."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.1.1",
    "jinja2>=3.1.6",
    "python-frontmatter>=1.1.0",
    "gradio>=5.44.1",
    "uvicorn>=0.35.0",
    "fastapi[standard]>=0.116.1",
    "requests>=2.28.0",
]
